+++ test/arx_terbilang_test.dart
@@ -4,8 +4,7 @@ import 'package:arx_terbilang/arx_terbilang.dart';
 
 void main() {
   test('Return string', () {
-    final terbilang = Terbilang();
-    expect(terbilang.createInt(5431), isA<String>());
-    expect(terbilang.createInt(0), isA<String>());
+    expect(Terbilang.angka(5431), isA<String>());
+    expect(Terbilang.angka(2), isA<String>());
   });
 }