+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-## 0.0.1
+## 0.0.3
 
-* TODO: Describe initial release.
+- Add quadrillion support, prefix and suffix parameter
+
+## 0.0.2
+
+- Add trim function to result string